@@ -14,6 +14,11 @@
     "@heroicons/react": "^2.1.1",
     "@hookform/resolvers": "^3.9.0",
     "@material-tailwind/react": "^2.1.9",
+    "@nextui-org/card": "^2.0.33",
+    "@nextui-org/react": "^2.4.6",
+    "@nextui-org/slider": "^2.2.15",
+    "@nextui-org/system": "^2.2.5",
+    "@nextui-org/theme": "^2.2.9",
     "@prisma/client": "^5.10.2",
     "@radix-ui/react-avatar": "^1.1.0",
     "@radix-ui/react-checkbox": "^1.1.1",
@@ -22,10 +27,13 @@
     "@radix-ui/react-dropdown-menu": "^2.1.1",
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-popover": "^1.1.1",
+    "@radix-ui/react-radio-group": "^1.2.0",
     "@radix-ui/react-scroll-area": "^1.1.0",
     "@radix-ui/react-select": "^2.1.1",
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-toggle": "^1.1.0",
+    "@radix-ui/react-toggle-group": "^1.1.0",
     "@reduxjs/toolkit": "^2.2.1",
     "@tanstack/react-table": "^8.19.3",
     "autoprefixer": "^10.4.19",
@@ -43,6 +51,7 @@
     "http-status-codes": "^2.3.0",
     "input-otp": "^1.2.4",
     "is-ci": "^3.0.1",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.408.0",
     "moment": "^2.30.1",
     "next": "14.1.3",
@@ -50,7 +59,7 @@
     "next-themes": "^0.2.1",
     "nodemailer": "^6.9.14",
     "paypal-rest-sdk": "^1.8.1",
-    "postcss": "^8.4.38",
+    "postcss": "^8.4.40",
     "razorpay": "^2.9.4",
     "react": "^18",
     "react-alice-carousel": "^2.9.1",
@@ -78,16 +87,16 @@
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
-    "@types/node": "^20.11.25",
+    "@types/node": "^20.14.12",
     "@types/nodemailer": "^6.4.15",
     "@types/paypal-rest-sdk": "^1.7.9",
-    "@types/react": "^18.2.64",
-    "@types/react-dom": "^18",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@types/react-js-pagination": "^3.0.7",
     "@types/react-slick": "^0.23.13",
     "prisma": "^5.10.2",
-    "tailwindcss": "^3.4.1",
+    "tailwindcss": "^3.4.7",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.2"
+    "typescript": "^5.5.4"
   }
 }
