@@ -18,8 +18,10 @@ import { reviewSubmit, getReviews } from "@/app/redux/slices/reviewSlice";
 import ProductPreview from '../../../../components/ProductPreview';
 import Image from "next/image";
 import { FaEdit } from 'react-icons/fa';
-
-
+import { DateRange } from "react-day-picker"
+import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group"
+import { Button } from "@/components/ui/button"
+import { StarIcon } from '@/components/index'
 
 export default function Checkout({ params }: { params: { product: string } }) {
     const productsList: any = useSelector((state: any) => state.categories.productsList);
@@ -40,6 +42,10 @@ export default function Checkout({ params }: { params: { product: string } }) {
     const orders: any[] = useAppSelector((state) => state?.orderReducer?.orders);
     const wishlist: any[] = useAppSelector((state) => state?.wishListReducer?.wishList);
     const [info, setInfo] = useState("description");
+    const today = new Date();
+    const oneMonthAgo = new Date(today);
+    oneMonthAgo.setMonth(today.getMonth() - 1);
+    const [date, setDate] = useState<DateRange | undefined>({ from: oneMonthAgo, to: today, })
 
     const isCurrentUserReviewed = useMemo(() => {
         return reviews.some((review: any) => review?.userId === session?.user?.id && review?.productId === product.id);
@@ -90,15 +96,22 @@ export default function Checkout({ params }: { params: { product: string } }) {
         session && product?.id &&
             (async () => {
                 const data = await dispatch(getReviews(product?.id));
-                const orders: any = await dispatch(getOrdersFunc());
+                const orders: any = await dispatch(getOrdersFunc({ page: 1, limit: 1000, search: "", from: date?.from?.toString(), to: date?.to?.toString(), slug: "getAll", }));
                 const payload = orders?.payload?.data
 
                 payload?.some((item: any) => {
-                    const purchaseFilter = item?.OrderItem?.some((item: any) => item?.productId === product?.id);
-                    if (purchaseFilter) {
-                        setIsPurchased(true);
-                        return true;
+
+                    if (item?.orderStatus === "COMPLETE") {
+                        console.log('item::: ', item);
+                        const purchaseFilter = item?.OrderItem?.some((item: any) => item?.productId === product?.id);
+
+                        if (purchaseFilter) {
+                            setIsPurchased(true);
+                            return true;
+                        }
                     }
+
+
                     return false;
                 });
             })();
@@ -128,147 +141,128 @@ export default function Checkout({ params }: { params: { product: string } }) {
 
     return (
         <>
-            <div className="grid grid-cols-2 w-full h-full mt-5">
-
-                <div className="carousel-container ">
-                    <Carousel showThumbs={false}>
-                        {images_video}
-                    </Carousel>
-                </div>
-
-                <div className="bg-white p-5 ">
-                    <div className="border-gray-400 pb-5  border-b-2">
-                        <div className="flex items-center gap-5">
-                            <p className="text-4xl font-medium">{product?.name}</p>
-                            <div className="text-xl font-light bg-green-900 text-white px-3 py-1 rounded-full">
-                                {product?.discountType === "PERCENTAGE" ? <span className="text-sm">{product?.discount}% OFF</span> :
-                                    <span className="text-sm">{product?.discount} ₹ OFF</span>
-                                }
-                            </div>
-                        </div>
-                        <div className="flex py-2 items-center gap-5">
-                            {product?.discountType === "PERCENTAGE" ? <span className="text-2xl">₹ {product?.price - product?.price * product?.discount / 100}</span> :
-                                <span className="text-2xl">₹ {product?.price - product?.discount} </span>
-                            }
-                            <p> {product?.price}</p>
-                        </div>
-                        <div className="flex items-center gap-2">
-                            <div className="flex gap-1 text-sm text-orange-900">
-                                <FaStar />
-                                <FaStar />
-                                <FaStar />
-                                <FaStar />
-                                <FaStar />
-                            </div>
-                            <p className="text-sm">(45)</p>
-                        </div>
+            <div className="max-w-6xl px-4 mx-auto py-6 border border-black">
+                <div className="grid md:grid-cols-2 gap-6 lg:gap-12 items-start">
+                    <div className="grid gap-4 md:gap-10 items-start">
+                        <img
+                            src="/placeholder.svg"
+                            alt="Product Image"
+                            width={600}
+                            height={300}
+                            className="aspect-[2/3] object-cover w-full border border-black h-[56vh] rounded-lg overflow-hidden"
+                        />
                     </div>
-                    <div className="py-5">
-                        <div className="font-medium flex items-center gap-3">
-                            <p>Availability:</p>
-                            <p className="text-green-300">In Stock</p>
-                        </div>
-                        <p className="py-5">
-                            Versatile, comfortable, and chic! Three words that describe Blake
-                            by Elyssi.
-                        </p>
-                        <div className="grid grid-cols-7 items-center gap-5">
-                            <p>Color</p>
-                            <div className="flex gap-1">
-                                <FaCircle />
-                                <FaCircle />
-                                <FaCircle />
-                                <FaCircle />
-                            </div>
-                        </div>{" "}
-                        <div className="grid grid-cols-7 items-center gap-5 py-5">
-                            <p>size</p>
-                            <div className="col-span-3">
-                                <input type="text" placeholder="small" className="w-full" />
-                            </div>
-                        </div>
-                        <div className="grid grid-cols-7 items-center gap-5 py-5">
-                            <p>Quantity</p>
-                            <div className="col-span-3">
-                                <input type="text" placeholder="small" className="w-full" />
-                            </div>
-                        </div>
-                        <div className="flex items-center gap-5">
-                            <button>Add to cart </button>
-                            <button> buy now</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
 
-            <div className="bg-white p-5">
-                <div className="flex items-center mb-5">
-                    <button
-                        className={`p-3 ${info === "description" ? "bg-gray-300" : ""}`}
-                        onClick={() => setInfo("description")}
-                    >
-                        Description
-                    </button>
-                    <button
-                        className={`p-3 ${info === "reviews" ? "bg-gray-300" : ""}`}
-                        onClick={() => setInfo("reviews")}
-                    >
-                        Reviews
-                    </button>
-                </div>
-                <div className="bg-gray-100 p-5 rounded-lg">
-                    {info === "description" && (
-                        <>
-                            <p>{product?.description}</p>
-                        </>
-                    )}
-                    {info === "reviews" && (
-                        <>
-                            {
-                                session && isPurchased && !isCurrentUserReviewed &&
-                                <div className="flex justify-end mb-4">
-                                    <button
-                                        onClick={() => setShowModal(true)}
-                                        className="bg-black text-white px-4 py-2 rounded-lg shadow-md hover:bg-gray-600 transition duration-300"
-                                    >
-                                        Add Review
-                                    </button>
+                    <div className="grid gap-4 md:gap-10 items-start">
+                        <div className="grid gap-2">
+                            <h1 className="font-bold text-3xl lg:text-4xl">Cozy Cotton Crew Neck Sweater</h1>
+                            <div className="flex items-center gap-4">
+                                <div className="flex items-center gap-0.5">
+                                    <StarIcon className="w-5 h-5 fill-primary" />
+                                    <StarIcon className="w-5 h-5 fill-primary" />
+                                    <StarIcon className="w-5 h-5 fill-primary" />
+                                    <StarIcon className="w-5 h-5 fill-muted stroke-muted-foreground" />
+                                    <StarIcon className="w-5 h-5 fill-muted stroke-muted-foreground" />
                                 </div>
-
-                            }
-                            <div className="max-h-96 overflow-y-auto">
-                                {reviews.map((review: any) => (
-                                    <div key={review.id} className="mb-2 p-4 border rounded-lg bg-white bg-opacity-75 shadow-lg flex items-center space-x-4">
-                                        <div className="w-10">
-                                            <Image src={`/users/${review?.user?.profile_pic}` || fallbackImage} alt={`Review ${review.id}`} width={100} height={100} className="w-full  rounded-lg" />
-                                        </div>
-                                        <div className="w-3/4">
-                                            <div className="flex items-center mb-1">
-                                                {[...Array(5)].map((star, index) => (
-                                                    <FaStar
-                                                        key={index}
-                                                        className={`mr-1 ${index < review?.rating ? "text-yellow-500" : "text-gray-300"}`}
-                                                    />
-                                                ))}
-                                            </div>
-                                            <p className="text-gray-800 mb-1">{review?.review}</p>
-                                            <p className="text-gray-500 text-sm">- {review?.user?.name}, {review?.user?.country}</p>
-                                        </div>
-                                        {
-                                            session.user.id === review?.user?.id &&
-                                            <div>
-                                                <button className='flex gap-2' onClick={() => { setShowModal(true); setNewReview({ rating: review?.rating, text: review?.review, }) }} >
-                                                    <FaEdit />
-                                                </button>
-                                            </div>
-                                        }
-                                    </div>
-                                ))}
+                                <div className="text-sm text-muted-foreground">(4.3 out of 5)</div>
                             </div>
-                        </>
-                    )}
+                            <div className="flex items-center gap-4">
+                                <div className="text-4xl font-bold">$59.99</div>
+                                <div className="text-sm text-muted-foreground line-through">$79.99</div>
+                                <div className="inline-block rounded-lg bg-primary px-3 py-1 text-sm text-primary-foreground">25% OFF</div>
+                            </div>
+                        </div>
+                        <div className="grid gap-4 text-sm leading-loose">
+                            <p>
+                                {/* Crafted from premium cotton, this cozy crew neck sweater is the perfect addition to your wardrobe. With a
+              classic design and comfortable fit, it's ideal for everyday wear or layering during the colder months. */}
+                            </p>
+                            <div className="grid gap-2">
+                                <div className="font-semibold">Size:</div>
+                                <div className="flex items-center gap-2">
+                                    <ToggleGroup type="single" defaultValue="m" variant="outline">
+                                        <ToggleGroupItem value="s">S</ToggleGroupItem>
+                                        <ToggleGroupItem value="m">M</ToggleGroupItem>
+                                        <ToggleGroupItem value="l">L</ToggleGroupItem>
+                                        <ToggleGroupItem value="xl">XL</ToggleGroupItem>
+                                    </ToggleGroup>
+                                </div>
+                            </div>
+                        </div>
+                        <Button size="lg">Add to Cart</Button>
+                        <Button size="lg">Buy Now</Button>
+                    </div>
+                </div>
+
+                <div className="grid  bg-white mt-2 p-5">
+                    <div className="flex items-center mb-5">
+                        <button
+                            className={`p-3 ${info === "description" ? "bg-gray-300" : ""}`}
+                            onClick={() => setInfo("description")}
+                        >
+                            Description
+                        </button>
+                        <button
+                            className={`p-3 ${info === "reviews" ? "bg-gray-300" : ""}`}
+                            onClick={() => setInfo("reviews")}
+                        >
+                            Reviews
+                        </button>
+                    </div>
+                    <div className="bg-gray-100 p-5 rounded-lg">
+                        {info === "description" && (
+                            <>
+                                <p>{product?.description}</p>
+                            </>
+                        )}
+                        {info === "reviews" && (
+                            <>
+                                {session && isPurchased && !isCurrentUserReviewed && (
+                                    <div className="flex justify-end mb-4">
+                                        <button
+                                            onClick={() => setShowModal(true)}
+                                            className="bg-black text-white px-4 py-2 rounded-lg shadow-md hover:bg-gray-600 transition duration-300"
+                                        >
+                                            Add Review
+                                        </button>
+                                    </div>
+                                )}
+                                <div className="max-h-96 overflow-y-auto">
+                                    {reviews.map((review: any) => (
+                                        <div key={review.id} className="mb-2 p-4 border rounded-lg bg-white bg-opacity-75 shadow-lg flex items-center space-x-4">
+                                            <div className="w-10">
+                                                <Image src={`/users/${review?.user?.profile_pic}` || fallbackImage} alt={`Review ${review.id}`} width={100} height={100} className="w-full rounded-lg" />
+                                            </div>
+                                            <div className="w-3/4">
+                                                <div className="flex items-center mb-1">
+                                                    {[...Array(5)].map((star, index) => (
+                                                        <FaStar
+                                                            key={index}
+                                                            className={`mr-1 ${index < review?.rating ? "text-yellow-500" : "text-gray-300"}`}
+                                                        />
+                                                    ))}
+                                                </div>
+                                                <p className="text-gray-800 mb-1">{review?.review}</p>
+                                                <p className="text-gray-500 text-sm">- {review?.user?.name}, {review?.user?.country}</p>
+                                            </div>
+                                            {session?.user.id === review?.user?.id && (
+                                                <div>
+                                                    <button className="flex gap-2" onClick={() => { setShowModal(true); setNewReview({ rating: review?.rating, text: review?.review }) }}>
+                                                        <FaEdit />
+                                                    </button>
+                                                </div>
+                                            )}
+                                        </div>
+                                    ))}
+                                </div>
+                            </>
+                        )}
+                    </div>
                 </div>
             </div>
+
+
+          
 
             <div className="p-5">
                 <div className="flex justify-center items-center uppercase  text-5xl pt-10 font-normal text- unica-one">
@@ -339,9 +333,213 @@ export default function Checkout({ params }: { params: { product: string } }) {
                     </div>
                 </div>
             )}
-            {/* <ProductPreview /> */}
-
         </>
     )
 }
+
+
+
+// export default function Checkout({ params }: { params: { product: string } }) {
+
+//     return (
+//         <>
+//             {/* <div className="grid grid-cols-2 w-full h-full mt-5">
+
+//                 <div className="carousel-container ">
+//                     <Carousel showThumbs={false}>
+//                         {images_video}
+//                     </Carousel>
+//                 </div>
+
+//                 <div className="bg-white p-5 ">
+//                     <div className="border-gray-400 pb-5  border-b-2">
+//                         <div className="flex items-center gap-5">
+//                             <p className="text-4xl font-medium">{product?.name}</p>
+//                             <div className="text-xl font-light bg-green-900 text-white px-3 py-1 rounded-full">
+//                                 {product?.discountType === "PERCENTAGE" ? <span className="text-sm">{product?.discount}% OFF</span> :
+//                                     <span className="text-sm">{product?.discount} ₹ OFF</span>
+//                                 }
+//                             </div>
+//                         </div>
+//                         <div className="flex py-2 items-center gap-5">
+//                             {product?.discountType === "PERCENTAGE" ? <span className="text-2xl">₹ {product?.price - product?.price * product?.discount / 100}</span> :
+//                                 <span className="text-2xl">₹ {product?.price - product?.discount} </span>
+//                             }
+//                             <p> {product?.price}</p>
+//                         </div>
+//                         <div className="flex items-center gap-2">
+//                             <div className="flex gap-1 text-sm text-orange-900">
+//                                 <FaStar />
+//                                 <FaStar />
+//                                 <FaStar />
+//                                 <FaStar />
+//                                 <FaStar />
+//                             </div>
+//                             <p className="text-sm">(45)</p>
+//                         </div>
+//                     </div>
+//                     <div className="py-5">
+//                         <div className="font-medium flex items-center gap-3">
+//                             <p>Availability:</p>
+//                             <p className="text-green-300">In Stock</p>
+//                         </div>
+//                         <p className="py-5">
+//                             Versatile, comfortable, and chic! Three words that describe Blake
+//                             by Elyssi.
+//                         </p>
+//                         <div className="grid grid-cols-7 items-center gap-5">
+//                             <p>Color</p>
+//                             <div className="flex gap-1">
+//                                 <FaCircle />
+//                                 <FaCircle />
+//                                 <FaCircle />
+//                                 <FaCircle />
+//                             </div>
+//                         </div>{" "}
+//                         <div className="grid grid-cols-7 items-center gap-5 py-5">
+//                             <p>size</p>
+//                             <div className="col-span-3">
+//                                 <input type="text" placeholder="small" className="w-full" />
+//                             </div>
+//                         </div>
+//                         <div className="grid grid-cols-7 items-center gap-5 py-5">
+//                             <p>Quantity</p>
+//                             <div className="col-span-3">
+//                                 <input type="text" placeholder="small" className="w-full" />
+//                             </div>
+//                         </div>
+//                         <div className="flex items-center gap-5">
+//                             <button>Add to cart </button>
+//                             <button> buy now</button>
+//                         </div>
+//                     </div>
+//                 </div>
+//             </div> */}
+
+{/* <div className="grid md:grid-cols-2 gap-6 lg:gap-12   border border-black  items-start max-w-6xl px-4 mx-auto py-6">
+<div className="grid gap-4 md:gap-10 items-start">
+    <img
+        src="/placeholder.svg"
+        alt="Product Image"
+        width={600}
+        height={300}
+        className="aspect-[2/3] object-cover  w-full border border-black h-[56vh] rounded-lg overflow-hidden"
+    />
+</div>
+
+<div className="grid gap-4 md:gap-10    items-start">
+    <div className="grid gap-2">
+        <h1 className="font-bold text-3xl lg:text-4xl">Cozy Cotton Crew Neck Sweater</h1>
+        <div className="flex items-center gap-4">
+            <div className="flex items-center gap-0.5">
+                <StarIcon className="w-5 h-5 fill-primary" />
+                <StarIcon className="w-5 h-5 fill-primary" />
+                <StarIcon className="w-5 h-5 fill-primary" />
+                <StarIcon className="w-5 h-5 fill-muted stroke-muted-foreground" />
+                <StarIcon className="w-5 h-5 fill-muted stroke-muted-foreground" />
+            </div>
+            <div className="text-sm text-muted-foreground">(4.3 out of 5)</div>
+        </div>
+        <div className="flex items-center gap-4">
+            <div className="text-4xl font-bold">$59.99</div>
+            <div className="text-sm text-muted-foreground line-through">$79.99</div>
+            <div className="inline-block rounded-lg bg-primary px-3 py-1 text-sm text-primary-foreground">25% OFF</div>
+        </div>
+    </div>
+    <div className="grid gap-4 text-sm leading-loose">
+        <p>
+            {/* Crafted from premium cotton, this cozy crew neck sweater is the perfect addition to your wardrobe. With a
+            classic design and comfortable fit, it's ideal for everyday wear or layering during the colder months. */}
+//         </p>
+//         <div className="grid gap-2">
+//             <div className="font-semibold">Size:</div>
+//             <div className="flex items-center gap-2">
+//                 <ToggleGroup type="single" defaultValue="m" variant="outline">
+//                     <ToggleGroupItem value="s">S</ToggleGroupItem>
+//                     <ToggleGroupItem value="m">M</ToggleGroupItem>
+//                     <ToggleGroupItem value="l">L</ToggleGroupItem>
+//                     <ToggleGroupItem value="xl">XL</ToggleGroupItem>
+//                 </ToggleGroup>
+//             </div>
+//         </div>
+//     </div>
+//     <Button size="lg">Add to Cart</Button>
+//     <Button size="lg">Buy Now</Button>
+// </div>
+// </div>
+
+{/* <div className="grid border border-black bg-white mt-2 p-5">
+<div className="flex items-center mb-5">
+    <button
+        className={`p-3 ${info === "description" ? "bg-gray-300" : ""}`}
+        onClick={() => setInfo("description")}
+    >
+        Description
+    </button>
+    <button
+        className={`p-3 ${info === "reviews" ? "bg-gray-300" : ""}`}
+        onClick={() => setInfo("reviews")}
+    >
+        Reviews
+    </button>
+</div>
+<div className="bg-gray-100 p-5 rounded-lg">
+    {info === "description" && (
+        <>
+            <p>{product?.description}</p>
+        </>
+    )}
+    {info === "reviews" && (
+        <>
+            {
+                session && isPurchased && !isCurrentUserReviewed &&
+                <div className="flex justify-end mb-4">
+                    <button
+                        onClick={() => setShowModal(true)}
+                        className="bg-black text-white px-4 py-2 rounded-lg shadow-md hover:bg-gray-600 transition duration-300"
+                    >
+                        Add Review
+                    </button>
+                </div>
+
+            }
+            <div className="max-h-96 overflow-y-auto">
+                {reviews.map((review: any) => (
+                    <div key={review.id} className="mb-2 p-4 border rounded-lg bg-white bg-opacity-75 shadow-lg flex items-center space-x-4">
+                        <div className="w-10">
+                            <Image src={`/users/${review?.user?.profile_pic}` || fallbackImage} alt={`Review ${review.id}`} width={100} height={100} className="w-full  rounded-lg" />
+                        </div>
+                        <div className="w-3/4">
+                            <div className="flex items-center mb-1">
+                                {[...Array(5)].map((star, index) => (
+                                    <FaStar
+                                        key={index}
+                                        className={`mr-1 ${index < review?.rating ? "text-yellow-500" : "text-gray-300"}`}
+                                    />
+                                ))}
+                            </div>
+                            <p className="text-gray-800 mb-1">{review?.review}</p>
+                            <p className="text-gray-500 text-sm">- {review?.user?.name}, {review?.user?.country}</p>
+                        </div>
+                        {
+                            session?.user.id === review?.user?.id &&
+                            <div>
+                                <button className='flex gap-2' onClick={() => { setShowModal(true); setNewReview({ rating: review?.rating, text: review?.review, }) }} >
+                                    <FaEdit />
+                                </button>
+                            </div>
+                        }
+                    </div>
+                ))}
+            </div>
+        </>
+    )}
+</div>
+</div> */}
+
+//           
+
+//         </>
+//     )
+// }
 
