@@ -1,35 +1,44 @@
+import React from "react";
 import type { Metadata } from "next";
+
 import "./globals.css";
 import { Providers } from "./redux/provider";
 import { AuthProvider } from '@/providers/Provider';
-import { Inter as FontSans } from "next/font/google"
-
+import { Inter } from 'next/font/google'
 import { cn } from "@/lib/utils"
 
-const fontSans = FontSans({
-  subsets: ["latin"],
-  variable: "--font-sans",
-})
+
+const fontHeading = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-heading',
+});
+
+const fontBody = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-body',
+});
+
+
 
 type RootLayoutProps = {
   children: React.ReactNode;
 };
 
-export default function RootLayout({ children }: RootLayoutProps) {
+export default function Layout({ children }: RootLayoutProps) {
   return (
-    <html lang="en" suppressHydrationWarning>
+    <html lang="en">
       <head>
         <meta charSet="utf-8" />
-        <link
-          rel="stylesheet"
-          href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap"
-        />
+      
         <link rel="icon" href="/favicon2.png" type="image/png" sizes="32x32" />
       </head>
       <body
         className={cn(
-          "min-h-screen bg-background font-sans antialiased",
-          fontSans.variable
+          'antialiased',
+          fontHeading.variable,
+          fontBody.variable
         )}
       >
         <Providers >
@@ -37,9 +46,7 @@ export default function RootLayout({ children }: RootLayoutProps) {
             {children}
           </AuthProvider>
         </Providers>
-
       </body>
     </html>
-  )
+  );
 }
-
