@@ -4,10 +4,19 @@ interface ThunkApiConfig {
     rejectWithValue: any;
 }
 
-export const getReturnOrdersFunc = createAsyncThunk('returnOrder/getReturnOrderFunc', async (_, thunkApiConfig: ThunkApiConfig) => {
+export const getReturnOrdersFunc = createAsyncThunk('returnOrder/getReturnOrderFunc', async ({ page, limit, search, from, to, slug, }: { page: number; limit: number; search: any; from: any, to: any, slug: string }, thunkApiConfig: ThunkApiConfig) => {
     const { rejectWithValue } = thunkApiConfig;
     try {
-        const response = await axios.get(`${process.env.NEXT_PUBLIC_API_URL}/return/returnOrder/getAll`)
+        const params = new URLSearchParams({
+            page: page.toString(),
+            limit: limit.toString(),
+            search: search.toString(),
+            from: from.toString(),
+            to: to.toString(),
+            slug: slug.toString()
+        }).toString();
+
+        const response = await axios.get(`${process.env.NEXT_PUBLIC_API_URL}/return/returnOrder?${params}`)
         return response.data;
     } catch (error) {
         const errorMessage = (error as Error).message || 'Unknown error occurred';
