@@ -1,7 +1,6 @@
 import Image from "next/image";
 import React from "react";
-import { FaStar, FaHeart } from "react-icons/fa6";
-import { FaRegHeart } from "react-icons/fa";
+import { FaStar, FaHeart, FaRegHeart } from "react-icons/fa";
 import { useAppSelector, useAppDispatch } from '@/app/redux/hooks';
 import { useSession } from "next-auth/react";
 import { errorToast, successToast } from "@/components/toster";
@@ -9,6 +8,10 @@ import { addToWishList } from "@/app/redux/slices/wishListSlice";
 import { isLoginModel } from '@/app/redux/slices/utilSlice';
 import { setOpenCart } from '@/app/redux/slices/utilSlice';
 import { actionTocartFunc } from '@/app/redux/slices/cartSclice';
+import Link from "next/link";
+import { Button } from "@/components/ui/button";
+import { Card, CardBody, CardFooter } from "@nextui-org/react";
+import { StarRating } from "../TopselectionCard/page";
 
 const Recentviewedcard = ({
   item,
@@ -18,106 +21,111 @@ const Recentviewedcard = ({
   wish: boolean
 }) => {
   const dispatch = useAppDispatch();
-  const { data: session, status }: any = useSession();
+  const { data: session }: any = useSession();
   const cart = useAppSelector((state) => state?.cartReducer?.cart?.CartItem) || [];
   const openCart = useAppSelector((state) => state?.utilReducer?.openCart);
 
-
-  const handelike = async () => {
+  const handleLike = async () => {
     if (session) {
-      dispatch(addToWishList({ productId: item?.id, }));
+      dispatch(addToWishList({ productId: item?.id }));
     } else {
       dispatch(isLoginModel(true));
     }
   };
 
   const addToCartFunction = async (id: string) => {
-    const payload = { productId: id, action: "add" }
-    const data = await dispatch(actionTocartFunc(payload))
+    const payload = { productId: id, action: "add" };
+    const data = await dispatch(actionTocartFunc(payload));
     if (data.payload.st) {
-      successToast(data?.payload.msg)
+      successToast(data?.payload.msg);
     } else {
-      errorToast(data.payload.msg)
+      errorToast(data.payload.msg);
     }
-  }
+  };
+
+  const buyNowFunction = async (id: string) => {
+    // Implement the buy now functionality here
+    // This could involve directly navigating to the checkout page with the selected item
+  };
 
   return (
-    <>
-      <div className="border-2 border-[#CFCFCF] shadow-xl mx-5 my-2 bg-white pb-5 w-72 h-100 flex flex-col justify-between">
-        <div>
-          {item?.isNew ? (
-            <div className="flex justify-between">
-              <div className="font-semibold text-base px-7 flex justify-center items-center bg-black text-white">
-                NEW
-              </div>
-              <div className="pr-5 pt-5">
-                <button onClick={handelike}>
-                  {wish ? (
-                    <FaHeart className="w-7 h-7 text-red-500" />
-                  ) : (
-                    <FaRegHeart className="w-7 h-7 " />
-                  )}
-                </button>
-
-              </div>
-            </div>
+    <Card shadow="lg" isPressable className="w-[250px] border border-gray-300 max-w-sm my-5  md:mx-0 mx-3">
+      <CardBody className="overflow-visible p-0 relative ">
+        {item?.isNew && (
+          <div className="absolute top-2 left-2 bg-primary px-2 py-1 rounded-md text-white text-xs font-medium">
+            New Arrival
+          </div>
+        )}
+        <Image
+          loading="lazy"
+          width={300}
+          height={300}
+          alt={item.title}
+          className="w-full object-cover h-64 md:h-72"
+          src={`/products/${item?.image[0]}`}
+        />
+        <button
+          className={`absolute top-2 right-2 text-xl ${wish ? 'text-red-500' : 'text-gray-500'}`}
+          style={{ zIndex: 20, pointerEvents: 'auto' }}
+          onClick={(e) => {
+            e.stopPropagation();
+            session ? dispatch(addToWishList({ productId: item?.id })) : dispatch(isLoginModel(true));
+          }}
+        >
+          {wish ? '♥' : '♡'}
+        </button>
+      </CardBody>
+      <CardFooter className="flex flex-col items-start p-2 space-y-2">
+        <b className="text-lg font-semibold">{item.title}</b>
+        <p className="text-sm text-gray-700">
+          {item?.description.split(' ').slice(0, 15).join(' ')}...
+          <Link href={`/preview/${item?.id}`} className="text-orange-500">
+            More
+          </Link>
+        </p>
+        <div className="flex flex-col sm:flex-row justify-between items-center w-full space-y-2 sm:space-y-0">
+          <StarRating rating={item?.avgRating || 5} />
+          <div className="flex items-baseline space-x-2">
+            <p className="text-default-500 text-sm font-bold">₹ {item.price}</p>
+            <p className="text-green-500 text-sm font-semibold">
+              {item?.discountType === "PERCENTAGE" ? (
+                <span>{item?.discount}% off</span>
+              ) : (
+                <span>{item?.discount} ₹ off</span>
+              )}
+            </p>
+          </div>
+        </div>
+        <div className="flex flex-col justify-center sm:flex-row gap-2 w-full">
+          <Button variant="outline" className="w-full sm:w-auto border-green-300 text-green-600 hover:bg-green-100">
+            <Link href={`/buy/${item.id}`}>
+              Buy Now
+            </Link>
+          </Button>
+          {session && cart?.find((cartItem: any) => cartItem?.productId === item?.id) ? (
+            <Button
+              variant="outline"
+              className="w-full sm:w-auto border-blue-300 text-blue-600 hover:bg-blue-50"
+              onClick={() => {
+                session ? dispatch(setOpenCart(!openCart)) : "";
+              }}
+            >
+              Open cart
+            </Button>
           ) : (
-            <div className="flex justify-end">
-              <div className="pr-5 pt-5">
-                <button onClick={handelike}>
-
-                  {wish ? (
-                    <FaHeart className="w-7 h-7 text-red-500" />
-                  ) : (
-                    <FaRegHeart className="w-7 h-7 " />
-                  )}
-                </button>
-              </div>
-            </div>
+            <Button
+              variant="outline"
+              className="w-full sm:w-auto border-blue-300 text-blue-600 hover:bg-blue-50"
+              onClick={() => {
+                session ? addToCartFunction(item?.id) : dispatch(isLoginModel(true));
+              }}
+            >
+              Add to cart
+            </Button>
           )}
         </div>
-        <div className="pt-5 flex justify-center items-center flex-grow">
-          <Image
-            className="border border-black rounded object-cover"
-            src={`/products/${item?.image}`}
-            alt="NO Image Found!"
-            width={200}
-            height={200}
-          />
-        </div>
-        <div className="px-5">
-          <p className="text-2xl font-normal pt-5 text-center">{item?.label}</p>
-          <p className="pt-3 text-base text-center flex justify-center items-center gap-3">
-            <FaStar />
-            <span className="truncate">{item?.discription}</span>
-          </p>
-        </div>
-        <div className=" justify-center items-center pt-3  relative transition group mx-2">
-          <button className=" py-3  border-black border poppins  text-base font-bold w-full">
-            ₹{item?.price}
-          </button>
-          {
-            session && cart?.find((cartItem: any) => cartItem?.productId === item?.id) ?
-              <button
-                className="py-3 absolute left-0 border-black border poppins  text-base font-bold opacity-0 group-hover:opacity-100  w-full bg-black text-white"
-                onClick={() => {
-                  session ? dispatch(setOpenCart(!openCart)) : ""
-                }}
-              >
-                Open cart
-              </button> :
-              <button
-                className="py-3 absolute left-0 border-black border poppins  text-base font-bold opacity-0 group-hover:opacity-100  w-full bg-black text-white"
-                onClick={() => {
-                  session ? addToCartFunction(item?.id) : dispatch(isLoginModel(true));
-                }}
-              >
-                Add to cart
-              </button>
-          }
-        </div>
-      </div>
-    </>
+      </CardFooter>
+    </Card>
   );
 };
 
