@@ -8,19 +8,19 @@ const Deals = () => {
       </div>
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 py-14 px-20  gap-10 mx-auto ">
         <div className="relative flex justify-center items-center">
-          <img src="/image/deals1.jpg" alt="" />
+          <img src="/image/chubbyCheeks/bestdeals-1.jpeg" alt="" />
           <div className="glassmorph absolute  bottom-4  md:bottom-8 py-1 px-9 md:py-3  md:px-[75px] md:text-xl  lg:py-5 lg:px-28 lg:text-2xl font-medium poppins">
             Min 60% Off
           </div>
         </div>
         <div className=" relative flex justify-center items-center">
-          <img src="/image/deals2.jpg" alt="" />{" "}
+          <img src="/image/chubbyCheeks/bestdeals-2.jpeg" alt="" />{" "}
           <div className="glassmorph absolute bottom-4  md:bottom-8 py-1 px-5 md:py-3  md:px-14 md:text-xl  lg:py-5 lg:px-28 lg:text-2xl font-medium poppins">
             Buy 2 Get 1 Free
           </div>
         </div>
         <div className="relative  flex justify-center items-center">
-          <img src="/image/deals3.jpg" alt="" />{" "}
+          <img src="/image/chubbyCheeks/bestdeals-4.jpeg" alt="" />{" "}
           <div className="glassmorph absolute  bottom-4  md:bottom-8 py-1 px-9 md:py-3  md:px-[75px] md:text-xl  lg:py-5 lg:px-28 lg:text-2xl font-medium poppins">
             Min 40% Off
           </div>
